backend/requirements.txt
fastapi>=0.121.0
uvicorn[standard]>=0.38.0
SQLAlchemy>=2.0.41
alembic==1.11.1
psycopg2-binary>=2.9.9
python-multipart==0.0.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.26.124
python-dotenv==1.1.0
PyJWT==2.8.0
celery==5.3.1
redis==4.5.5
aiofiles==23.1.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
typing-extensions>=4.14.1
paddleocr>=2.7.0.3
pillow>=10.4.0
numpy>=1.26.4
onnxruntime>=1.18.1
insightface>=0.7.3
